git-svn-id: http://www.wesay.org/code/WeSay/trunk@480 b25276a9-ce15-0410-9281-991c76e3d784
committer: hatton <hatton@b25276a9-ce15-0410-9281-991c76e3d784>

+++ src/LexicalModel.Tests/Db4o-Specific/BaseDb4oSpecificTests.cs
@@ -22,10 +22,22 @@ namespace WeSay.LexicalModel.Tests
 		[TearDown]
 		public void TearDown()
 		{
-			this._entriesList.Dispose();
-		   this._dataSource.Dispose();
-			_recordListManager.Dispose();
-			File.Delete(_filePath);
+			if (_entriesList != null)
+			{
+				this._entriesList.Dispose();
+			}
+			if (_dataSource != null)
+			{
+				this._dataSource.Dispose();
+			}
+			if (_recordListManager != null)
+			{
+				_recordListManager.Dispose();
+			}
+			if (_filePath != "")
+			{
+				File.Delete(_filePath);
+			}
 		}
 
 		protected LexEntry CycleDatabase()
